+++ src/ScriptEditor/NetscriptDefinitions.d.ts
@@ -2033,26 +2033,6 @@ export interface Singularity {
   getAugmentationPrereq(augName: string): string[];
 
   /**
-   * Get the price and reputation of an augmentation.
-   * @deprecated use getAugmentationPrice getAugmentationRepReq
-   * @remarks
-   * RAM cost: 5 GB * 16/4/1
-   *
-   *
-   * This function returns an array with two elements that gives the cost for
-   * the specified Augmentation. The first element in the returned array is the
-   * reputation requirement of the Augmentation, and the second element is the
-   * money cost.
-   *
-   * If an invalid Augmentation name is passed in for the augName argument, this
-   * function will return the array [-1, -1].
-   *
-   * @param augName - Name of Augmentation.
-   * @returns Array with first element as a reputation requirement and second element as the money cost.
-   */
-  getAugmentationCost(augName: string): [number, number];
-
-  /**
    * Get price of an augmentation.
    * @remarks
    * RAM cost: 2.5 GB * 16/4/1
@@ -5915,33 +5895,6 @@ export interface NS {
    * @returns Base security level of the target server.
    */
   getServerBaseSecurityLevel(host: string): number;
-
-  /**
-   * @deprecated use getServerMaxRam / getServerUsedRam
-   * @remarks
-   * RAM cost: 0.1 GB
-   *
-   * Returns an array with two elements that gives information about a server’s memory (RAM).
-   * The first element in the array is the amount of RAM that the server has total (in GB).
-   * The second element in the array is the amount of RAM that is currently being used on
-   * the server (in GB).
-   *
-   * @example
-   * ```ts
-   * // NS1:
-   * var serverRam = getServerRam("helios");
-   * var totalRam = serverRam[0];
-   * var ramUsed = serverRam[1];
-   * ```
-   * @example
-   * ```ts
-   * // NS2:
-   * const [totalRam, ramUsed] = ns.getServerRam("helios");
-   * ```
-   * @param host - Host of target server.
-   * @returns Array with total and used memory on the specified server, in GB.
-   */
-  getServerRam(host: string): [number, number];
 
   /**
    * Get the maximum amount of RAM on a server.